refactor(web): share HTML tag scanning
Use one private reader for visibility filtering and rendering while preserving
their current malformed-input policies and caller-owned traversal. Skip quoted
spans without losing raw-text recovery. Readability and public APIs are unchanged.

Related: #141709
